requirements.txt
# CyberJustice Pedagogical Tutor Dependencies

# Core Flask dependencies
Flask==3.1.2
Flask-CORS==6.0.1
gunicorn==21.2.0

# LangChain and AI dependencies
langchain==0.3.27
langchain-groq==0.3.7
langchain-community==0.3.29
langchain-huggingface==0.3.1
langchain-core==0.3.75

# Vector database and embeddings
faiss-cpu==1.12.0
sentence-transformers==5.1.0

# Machine Learning and AI
torch==2.8.0
transformers==4.56.1

# Utilities
python-dotenv==1.1.1
beautifulsoup4==4.13.5

# HTTP and web scraping
requests==2.32.5
httpx==0.28.1

# Data processing
PyYAML==6.0.2
pydantic==2.11.7
numpy==2.3.2

# Additional dependencies
groq==0.31.1
tenacity==9.1.2
packaging==25.0
scikit-learn==1.7.1

# Note: This configuration supports the CyberJustice Pedagogical Tutor
# which uses Groq's qwen/qwen3-32b model for scaffolding and reflective learning
